+++ libraries/matrix/test/src/main/kotlin/io/element/android/libraries/matrix/test/room/FakeMatrixRoom.kt
@@ -387,7 +387,7 @@ class FakeMatrixRoom(
     override suspend fun sendVoiceMessage(
         file: File,
         audioInfo: AudioInfo,
-        waveform: List<Int>,
+        waveform: List<Float>,
         progressCallback: ProgressCallback?
     ): Result<MediaUploadHandler> = fakeSendMedia(progressCallback)
